Move deadline struct
