+++ pageserver/src/tenant_jobs/mod.rs
@@ -1,3 +1,4 @@
 pub mod compaction;
 pub mod gc;
 pub mod worker;
+pub mod deadline;
